Add argp parsing to validate command and add malloc check to previous commit
This commit is in accordance with the previous commit. It swaps the previously self implemented arg parsing with the argp package. It is prooving to give much more pleasant help and usage info and make more readable parsing logic. This commit also adds a malloc mem check to read, valicate and write source files.