Add autogenerated help/usage info for specific backends
This commit removes the hardcoded help/usage message for the 'secvarctl --help/usage' command. This will allow for different backends to easily define what commands are used and use there own descriptions. A new attribute to the command function was added, short_desc is a short description of what the command does. Also, temporarily the help and usage message for secvarctl are the same