Rename validate and ensure it is backend independent
This commit renames edk-svc-validate to general-validate. This is because the validate functions work with both efi and power secure variable files. Files generated with the latest efitools are able to be parsed, hopefully meaning that they are valid EFI secure boot files. Further testing should be done to ensure that they are able to be used in the secure variable management process. The only difference between an auth file generated with secvarctl and efitools is the PKCS7 containing extra header data versus only Signed Data. Luckily, our pkcs7 parser accepts both.