edk2-svc-verify.c: Do not add empty secvars to variable bank
Previously, if an empty (size 0) secvar existed in the sysfs, it was added to the current variable bank (see `setupBanks()`). This leads to errors in `validateBanks` because it calls `validateESL` on all entries in the current variable bank. `validateESL` returns an error if there is no data in the esl. This means that if there is no data in the current KEK, the verify function will return early on an error that the contained data in KEK/data is not a valid ESL. This commit, omits empty secvars from the current list of variables since an empty secvar is a common thing in Secure Boot and should not be treated as an error.

Signed-off-by: Nick Child <nick.child@ibm.com>