runTests.py: add test for empty secvar in verify command
This commit tests the implementation of the previous commit, which enabled `secvarctl verify` to not fail if an empty secvar exists. The test sets the size of the KEK variable to 0. Then a verify commmand is ran with an auth file for the db that is signed by the PK, previously this command would fail since not all of the current ESL's would validate. In fact, this test should pass since an empty KEK is entirely acceptable.

Signed-off-by: Nick Child <nick.child@ibm.com>